.bimi/: Add company logo in SVG form for email BIMI
A BIMI Record is a type of DNS Record used to display a company's logo
inside an email inbox if the email is trustworthy. Brand Indicator
Message Identification (BIMI) Records are an industry-wide effort to use
brand logos as indicators to help email recipients recognize and avoid
fraudulent messages.

Using BIMI requires ensuring DMARC authentication is set up on the domain.